Break out the docker service configuration into a separate script; move the vg check further down and not exit entire script, push/pop the rpm build and install; export env vars for the service config, print docker service and env at end.

+++ data/docker-storage-setup-strap.sh
@@ -2,11 +2,6 @@
 
 [ -e docker-storage-setup-env.sh ] && . ./docker-storage-setup-env.sh
 
-if vgdisplay | grep dockervg; then
-  echo 'Looks like the dockervg volume group already exists, skipping.'
-  exit 0
-fi
-
 # this is recommended but an up-to-date image should be fine
 [[ "$YUM_UPDATE" -eq 1 ]] && yum -y update
 
@@ -26,32 +21,33 @@ fi
 # install deps
 yum -y install rpm-build tar curl
 
-# download docker-storage-setup
-cd /tmp
-rm -Rf /tmp/docker-storage-setup*
-curl -L https://github.com/projectatomic/docker-storage-setup/archive/master.tar.gz > \
-  docker-storage-setup.tar.gz
-tar zxvf docker-storage-setup.tar.gz
-cd docker-storage-setup-master
+pushd /tmp
+  # download docker-storage-setup
+  rm -Rf /tmp/docker-storage-setup*
+  curl -L https://github.com/projectatomic/docker-storage-setup/archive/master.tar.gz > \
+    docker-storage-setup.tar.gz
+  tar zxvf docker-storage-setup.tar.gz
+  cd docker-storage-setup-master
 
-# temp hack (don't check VG_EXISTS)
-sed -i'' '315,318 s/^/#/' docker-storage-setup.sh
+  # temp hack (don't check VG_EXISTS)
+  sed -i'' '315,318 s/^/#/' docker-storage-setup.sh
 
-# turn on bash -x
-[[ "$DOCKER_STORAGE_DEBUG" -eq 1 ]] && \
-  sed -i 's%#!/bin/bash%#!/bin/bash -x%g' docker-storage-setup.sh
+  # turn on bash -x
+  [[ "$DOCKER_STORAGE_DEBUG" -eq 1 ]] && \
+    sed -i 's%#!/bin/bash%#!/bin/bash -x%g' docker-storage-setup.sh
 
-# build and install the docker-storage-setup rpm
-mkdir -p /root/rpmbuild/SOURCES
-rm -Rf /root/rpmbuild/SOURCES/*
-cp docker-storage-setup.sh \
-  docker-storage-setup.service \
-  docker-storage-setup.conf \
-  docker-storage-setup-override.conf \
-  libdss.sh \
-    /root/rpmbuild/SOURCES/
-rpmbuild -v -bb --clean docker-storage-setup.spec
-rpm -iv --force --replacepkgs /root/rpmbuild/RPMS/x86_64/*.rpm
+  # build and install the docker-storage-setup rpm
+  mkdir -p /root/rpmbuild/SOURCES
+  rm -Rf /root/rpmbuild/SOURCES/*
+  cp docker-storage-setup.sh \
+    docker-storage-setup.service \
+    docker-storage-setup.conf \
+    docker-storage-setup-override.conf \
+    libdss.sh \
+      /root/rpmbuild/SOURCES/
+  rpmbuild -v -bb --clean docker-storage-setup.spec
+  rpm -iv --force --replacepkgs /root/rpmbuild/RPMS/x86_64/*.rpm
+popd
 
 # save the overrides config
 touch /etc/sysconfig/docker-storage-setup
@@ -75,37 +71,42 @@ do
       echo "$i=${!i}" >> /etc/sysconfig/docker-storage-setup
 done
 
-# docker should already be enabled, but to be sure
-systemctl enable docker
+if vgdisplay | grep dockervg; then
+  echo 'Looks like the dockervg volume group already exists, skipping.'
+else
+  # ensure docker is stopped
+  systemctl stop docker
 
-# ensure docker is stopped
-systemctl stop docker
+  # start the docker-storage-setup service to reconfigure the disk
+  # systemctl start docker-storage-setup
+  #   or, command directly:
+  /usr/bin/docker-storage-setup
 
-# start the docker-storage-setup service to reconfigure the disk
-# systemctl start docker-storage-setup
-#   or, command directly:
-/usr/bin/docker-storage-setup
+  # optional, not required
+  # rm -Rf /var/lib/docker/*
+fi
 
-touch /etc/sysconfig/docker /etc/sysconfig/docker-storage /etc/sysconfig/docker-network
-mkdir -p /etc/systemd/system/docker.service.d
-cat <<'EOF'> /etc/systemd/system/docker.service.d/10-parameterized.conf
-[Service]
-EnvironmentFile=-/etc/sysconfig/docker
-EnvironmentFile=-/etc/sysconfig/docker-storage
-EnvironmentFile=-/etc/sysconfig/docker-network
-ExecStart=
-ExecStart=/usr/bin/docker daemon $OPTIONS \
-          $DOCKER_STORAGE_OPTIONS \
-          $DOCKER_NETWORK_OPTIONS \
-          $BLOCK_REGISTRY \
-          $INSECURE_REGISTRY
-EOF
+# export possible environment variables that can be used in service configuration
+export HTTP_PROXY
+export NO_PROXY
+export OPTIONS
+export DOCKER_STORAGE_OPTIONS
+export DOCKER_NETWORK_OPTIONS
+export BLOCK_REGISTRY
+export INSECURE_REGISTRY
 
-systemctl daemon-reload
-# rm -Rf /var/lib/docker/*
-systemctl start docker
+# configure the system service
+./docker-setup-service.sh
 
-# finally, print the docker info
+# finally, (re)start the docker service
+systemctl restart docker
+
+# print the docker info
 docker info
+
+# print the systemd docker service env
+systemctl show docker --property Environment
+
+systemctl status docker
 
 echo 'Done.'
